requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pypdf>=4.2.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
langchain-text-splitters>=0.2.1
streamlit>=1.36.0
tqdm>=4.66.0
